+++ tests/spcs/test_image_repository.py
@@ -142,15 +142,19 @@ def test_create_repository_already_exists(mock_handle, mock_execute):
 
 
 @patch(EXECUTE_QUERY)
-def test_deploy(mock_execute_query, runner, project_directory):
+def test_create_from_project_definition(
+    mock_execute_query, runner, project_directory, mock_cursor, os_agnostic_snapshot
+):
+    mock_execute_query.return_value = mock_cursor(
+        rows=[["Image Repository TEST_IMAGE_REPOSITORY successfully created."]],
+        columns=["status"],
+    )
+
     with project_directory("spcs_image_repository"):
-        result = runner.invoke(["spcs", "image-repository", "deploy"])
+        result = runner.invoke(["spcs", "image-repository", "create"])
 
         assert result.exit_code == 0, result.output
-        assert (
-            "Image repository 'test_image_repository' successfully deployed."
-            in result.output
-        )
+        assert result.output == os_agnostic_snapshot
         mock_execute_query.assert_called_once_with(
             "create image repository test_image_repository",
             name="test_image_repository",
@@ -158,17 +162,22 @@ def test_deploy(mock_execute_query, runner, project_directory):
 
 
 @patch(EXECUTE_QUERY)
-def test_deploy_replace(mock_execute_query, runner, project_directory):
+def test_create_from_project_definition_replace(
+    mock_execute_query, runner, project_directory, mock_cursor, os_agnostic_snapshot
+):
     image_repository_name = "test_image_repository"
+    mock_execute_query.return_value = mock_cursor(
+        rows=[
+            [f"Image Repository {image_repository_name.upper()} successfully created."]
+        ],
+        columns=["status"],
+    )
 
     with project_directory("spcs_image_repository"):
-        result = runner.invoke(["spcs", "image-repository", "deploy", "--replace"])
+        result = runner.invoke(["spcs", "image-repository", "create", "--replace"])
 
         assert result.exit_code == 0, result.output
-        assert (
-            f"Image repository '{image_repository_name}' successfully deployed."
-            in result.output
-        )
+        assert result.output == os_agnostic_snapshot
         mock_execute_query.assert_called_once_with(
             "create or replace image repository test_image_repository",
             name="test_image_repository",
@@ -176,7 +185,7 @@ def test_deploy_replace(mock_execute_query, runner, project_directory):
 
 
 @patch(EXECUTE_QUERY)
-def test_deploy_image_repository_already_exists(
+def test_create_from_project_definition_image_repository_already_exists(
     mock_execute_query, runner, project_directory
 ):
     mock_execute_query.side_effect = ProgrammingError(
@@ -184,7 +193,7 @@ def test_deploy_image_repository_already_exists(
     )
 
     with project_directory("spcs_image_repository"):
-        result = runner.invoke(["spcs", "image-repository", "deploy"])
+        result = runner.invoke(["spcs", "image-repository", "create"])
 
         assert result.exit_code == 1, result.output
         assert (
@@ -193,50 +202,56 @@ def test_deploy_image_repository_already_exists(
         )
 
 
-def test_deploy_no_image_repository(runner, project_directory):
+def test_create_from_project_definition_no_image_repository(runner, project_directory):
     with project_directory("empty_project"):
-        result = runner.invoke(["spcs", "image-repository", "deploy"])
+        result = runner.invoke(["spcs", "image-repository", "create"])
 
         assert result.exit_code == 1, result.output
         assert "No image repository project definition found in" in result.output
 
 
-def test_deploy_not_existing_entity_id(runner, project_directory):
+def test_create_from_project_definition_not_existing_entity_id(
+    runner, project_directory
+):
     with project_directory("spcs_image_repository"):
         result = runner.invoke(
-            ["spcs", "image-repository", "deploy", "not-existing-entity-id"]
+            ["spcs", "image-repository", "create", "not_existing_entity_id"]
         )
 
         assert result.exit_code == 2, result.output
         assert (
-            "No 'not-existing-entity-id' entity in project definition file."
+            "No 'not_existing_entity_id' entity in project definition file."
             in result.output
         )
 
 
 @patch(EXECUTE_QUERY)
-def test_deploy_multiple_image_repositories_with_entity_id(
-    mock_execute_query, runner, project_directory
+def test_create_from_project_definition_multiple_image_repositories_with_entity_id(
+    mock_execute_query, runner, project_directory, mock_cursor, os_agnostic_snapshot
 ):
+    mock_execute_query.return_value = mock_cursor(
+        rows=[["Image Repository TEST_IMAGE_REPOSITORY successfully created."]],
+        columns=["status"],
+    )
+
     with project_directory("spcs_multiple_image_repositories"):
         result = runner.invoke(
-            ["spcs", "image-repository", "deploy", "test_image_repository"]
+            ["spcs", "image-repository", "create", "test_image_repository"]
         )
 
         assert result.exit_code == 0, result.output
-        assert (
-            "Image repository 'test_image_repository' successfully deployed."
-            in result.output
-        )
+        assert result.output == os_agnostic_snapshot
         mock_execute_query.assert_called_once_with(
             "create image repository test_image_repository",
             name="test_image_repository",
         )
 
 
-def test_deploy_multiple_image_repositories(runner, project_directory):
+def test_create_from_project_definition_multiple_image_repositories(
+    runner, project_directory
+):
     with project_directory("spcs_multiple_image_repositories"):
-        result = runner.invoke(["spcs", "image-repository", "deploy"])
+        result = runner.invoke(["spcs", "image-repository", "create"])
 
         assert result.exit_code == 2, result.output
         assert (